feat: keep floating pane open on non-zero exit when close_on_exit is set
Propagate process exit code through the entire callback chain so that
close_on_exit only auto-closes the pane on success (exit code 0).
On abnormal exit the pane stays open for the user to inspect errors.

+++ TongYou/Terminal/TerminalControlling.swift
@@ -79,7 +79,7 @@ protocol TerminalControlling: AnyObject {
     // MARK: - Callbacks
 
     var onNeedsDisplay: (() -> Void)? { get set }
-    var onProcessExited: (() -> Void)? { get set }
+    var onProcessExited: ((Int32) -> Void)? { get set }
     var onTitleChanged: ((String) -> Void)? { get set }
     var onPaneNotification: ((String, String) -> Void)? { get set }
 }
