+++ TongYouTests/AltMouseSelectionTests.swift
@@ -15,7 +15,7 @@ private final class MockMouseController: TerminalControlling {
     var currentWorkingDirectory: String?
     var foregroundProcessName: String?
     var onNeedsDisplay: (() -> Void)?
-    var onProcessExited: (() -> Void)?
+    var onProcessExited: ((Int32) -> Void)?
     var onTitleChanged: ((String) -> Void)?
     var onPaneNotification: ((String, String) -> Void)?
